+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.2.0 (In development)
 
+- New API `NBTCodec#writeRegion(Path, ChunkRegion)` and `NBTCodec#writeRegion(Path, ChunkRegion, ExternalChunkAccessor)`
+  for writing chunk regions to files.
+
 ## 0.1.0 (2026-03-16)
 
 - Initial release